App.tsx를 페이지와 컴포넌트로 분리
