+++ src/App.tsx
@@ -1,53 +1,14 @@
 import React, { useState } from 'react';
-import logo from './logo.svg';
 import './App.css';
 import {
   Routes, Route, Outlet, Link, 
-  Navigate, useLocation, useNavigate,
 } from "react-router-dom";
 import Slider from './Components/Slider';
-import { useAuth } from './Contexts/Auth';
-import useString from './Hooks/StringState';
+import SigninPage from './Pages/SigninPage';
+import SignupPage from './Pages/SignupPage';
+import RequireAuth from './Components/RequireAuth';
 
 function App() {
-  let auth = useAuth();
-  let location = useLocation();
-  let navigate = useNavigate();
-
-  let from = location.state?.from?.pathname || "/";
-
-  const [username, setUsername] = useString();
-  const [password, setPassword] = useString();
-  const [message, setMessage] = useState<string>("");
-
-  const handleLogin = () => {
-    if (auth.signin(username, password)) {
-      reset();
-      navigate(from, { replace: true });
-    } else {
-      setMessage("유저 이름이나 비밀번호를 확인하세요.");
-    }
-  }
-
-  const handleSignup = () => {
-    if (username.length === 0) {
-      setMessage("유저 이름을 입력해주세요");
-    } else if (password.length === 0) {
-      setMessage("비밀번호를 입력해주세요");
-    } else if (auth.signup(username, password)) {
-      reset();
-      navigate(from, { replace: true });
-    } else {
-      setMessage("가입에 실패했습니다.");
-    }
-  }
-
-  const reset = () => {
-    setUsername("");
-    setPassword("");
-    setMessage("");
-  }
-  
   return (
     <>
       <LinkPage />
@@ -56,25 +17,9 @@ function App() {
         <Route path="/" element={<Slider />}>
         </Route>
         {/* login */}
-        <Route path="login" element={
-          <div>
-            login<br/>
-            유저 이름 <input className="border-2 border-black" type="text" id="username" name="username" value={username} onChange={e => setUsername(e.target.value)} required /><br/>
-            비밀번호 <input className="border-2 border-black" type="password" id="password" name="password" value={password} onChange={e => setPassword(e.target.value)} required /><br/>
-            <button className="border-2 border-black" onClick={() => handleLogin()}>로그인하기</button><br/>
-            {message}
-          </div>
-        } />
+        <Route path="login" element={<SigninPage />} />
         {/* sign up */}
-        <Route path="signup" element={
-          <div>
-            signup<br/>
-            유저 이름 <input className="border-2 border-black" type="text" id="username" name="username" value={username} onChange={e => setUsername(e.target.value)} required /><br/>
-            비밀번호 <input className="border-2 border-black" type="password" id="password" name="password" value={password} onChange={e => setPassword(e.target.value)} required /><br/>
-            <button className="border-2 border-black" onClick={() => handleSignup()}>가입하기</button><br/>
-            {message}
-          </div>} 
-        />
+        <Route path="signup" element={<SignupPage />} />
         {/* article */}
         <Route path="article">
           {/* article page */}
@@ -128,17 +73,6 @@ function LinkPage() {
       <Outlet />
     </div>
   );
-}
-
-function RequireAuth({ children }: { children: JSX.Element }) {
-  let auth = useAuth();
-  let location = useLocation();
-
-  if (!auth.user) {
-    return <Navigate to="/login" state={{ from: location }} />;
-  }
-
-  return children;
 }
 
 export default App;
